\name{rmixpanel-package}
\alias{rmixpanel-package}
\alias{rmixpanel}
\docType{package}
\title{
\packageTitle{rmixpanel}
}
\description{
\packageDescription{rmixpanel}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rmixpanel}
\packageIndices{rmixpanel}

}
\author{
\packageAuthor{rmixpanel}

Maintainer: \packageMaintainer{rmixpanel}
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
