\name{RMixpanel-package}
\alias{RMixpanel-package}
\alias{RMixpanel}
\docType{package}

\title{
\packageTitle{RMixpanel}
}

\description{
\packageDescription{RMixpanel}
}

\details{
The DESCRIPTION file:
\packageDESCRIPTION{RMixpanel}
\packageIndices{RMixpanel}
The package provides access to the Mixpanel API. It allows to read raw data as well as to retrieve retention, addiction and segmentation data.
}

\author{
\packageAuthor{RMixpanel}
Maintainer: \packageMaintainer{RMixpanel}
}

\references{
\url{https://mixpanel.com/docs/api-documentation/data-export-api}
}

\keyword{ package }
\keyword{ mixpanel }
