\name{rmixpanel-package}
\alias{rmixpanel-package}
\alias{rmixpanel}
\docType{package}
\title{
\packageTitle{rmixpanel}
}
\description{
\packageDescription{rmixpanel}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rmixpanel}
\packageIndices{rmixpanel}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rmixpanel}

Maintainer: \packageMaintainer{rmixpanel}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
